Rename figma component files to use kebab case
