+++ src/components/figma-property-preview.js
