+++ src/components/figma-property-table.js
